Use Bitmap GC from Ruby 2.0
Instead of marking inside of the object, we should keep a bitmap that
we can use to mark instead. By doing this, we do not dirty every object
that we see. The old way caused copy-on-write to occur during every mark
and this resulted in excessive memory consumption for CNUAPP when
forking workers via unicorn.

Reference for backport:
  https://github.com/funny-falcon/ruby/compare/p385...p385_backport_gc